Mouse new window support
