+++ RevitLookup/Services/Contracts/ILookupService.cs
@@ -28,6 +28,7 @@ public interface ILookupService : ILookupServiceDependsStage, ILookupServiceShow
 {
     ILookupServiceDependsStage Snoop(SnoopableType snoopableType);
     ILookupServiceDependsStage Snoop(SnoopableObject snoopableObject);
+    ILookupServiceDependsStage Snoop(IReadOnlyCollection<SnoopableObject> snoopableObjects);
     new ILookupServiceShowStage DependsOn(IServiceProvider provider);
     new ILookupServiceExecuteStage Show<T>() where T : Page;
 }
