+++ RevitLookup/Views/Pages/SnoopView.xaml
@@ -1,4 +1,4 @@
-<pages:SnoopViewBase
+<abstraction:SnoopViewBase
     x:Class="RevitLookup.Views.Pages.SnoopView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
@@ -8,6 +8,7 @@
     xmlns:pages="clr-namespace:RevitLookup.Views.Pages"
     xmlns:objects="clr-namespace:RevitLookup.Core.Objects"
     xmlns:converters="clr-namespace:RevitLookup.Views.Converters"
+    xmlns:abstraction="clr-namespace:RevitLookup.Views.Pages.Abstraction"
     mc:Ignorable="d"
     d:DesignHeight="500"
     d:DesignWidth="700"
@@ -72,11 +73,13 @@
                     VirtualizingStackPanel.IsVirtualizing="True"
                     ItemsSource="{Binding
                         ViewModel.FilteredSnoopableObjects,
-                        Converter={converters:TreeViewGroupConverter}}">
+                        Converter={converters:TreeViewGroupConverter}}"
+                    MouseMove="OnPresenterCursorInteracted">
                     <rl:TreeView.ItemTemplate>
                         <HierarchicalDataTemplate
                             ItemsSource="{Binding Items, Mode=OneTime}">
                             <TextBlock
+                                Loaded="OnHeaderLoaded"
                                 Text="{Binding Name, Mode=OneTime}" />
                             <HierarchicalDataTemplate.ItemTemplate>
                                 <DataTemplate>
@@ -99,13 +102,15 @@
                 ColumnHeaderHeight="30"
                 HeadersVisibility="Column"
                 AutoGenerateColumns="False"
+                SelectionMode="Single"
                 CanUserAddRows="False"
                 CanUserDeleteRows="False"
                 CanUserResizeRows="False"
                 CanUserSortColumns="False"
                 CanUserReorderColumns="False"
                 HorizontalScrollBarVisibility="Disabled"
-                ItemsSource="{Binding ViewModel.FilteredSnoopableData}">
+                ItemsSource="{Binding ViewModel.FilteredSnoopableData}"
+                MouseMove="OnPresenterCursorInteracted">
                 <rl:DataGrid.Columns>
                     <DataGridTemplateColumn
                         Width="*"
@@ -154,7 +159,7 @@
                         BasedOn="{StaticResource DefaultDataGridRowStyle}">
                         <EventSetter
                             Event="PreviewMouseLeftButtonUp"
-                            Handler="OnGridMouseLeftButtonUp" />
+                            Handler="OnGridRowClicked" />
                         <EventSetter
                             Event="Loaded"
                             Handler="OnRowLoaded" />
@@ -249,4 +254,4 @@
             </rl:DataGrid>
         </Grid>
     </Grid>
-</pages:SnoopViewBase>
+</abstraction:SnoopViewBase>
